backstage_templates_api/requirements.txt
fastapi
uvicorn
httpx
pydantic
pyyaml
typing-extensions
python-dotenv 
boto3